On-line Shop: registration and login steps were added, new table users in db was created, security service was implemented but should be updated, password storage should be updated

+++ src/main/java/com/luxcampus/shopOnline/web/servlet/DeleteRequestServlet.java
@@ -2,6 +2,7 @@ package com.luxcampus.shopOnline.web.servlet;
 
 import com.luxcampus.shopOnline.entity.Product;
 import com.luxcampus.shopOnline.service.ProductService;
+import com.luxcampus.shopOnline.service.SecurityService;
 import com.luxcampus.shopOnline.web.util.PageGenerator;
 
 import javax.servlet.ServletException;
@@ -9,13 +10,18 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.List;
 import java.util.Map;
 
 public class DeleteRequestServlet extends HttpServlet {
     private ProductService productService;
+    private List<String> userTokens;
+    private SecurityService securityService;
 
-    public DeleteRequestServlet(ProductService productService) {
+    public DeleteRequestServlet(ProductService productService, List<String> userTokens, SecurityService securityService) {
         this.productService = productService;
+        this.userTokens = userTokens;
+        this.securityService = securityService;
     }
 
     @Override
@@ -27,17 +33,21 @@ public class DeleteRequestServlet extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {
-        try {
-            int id = Integer.parseInt(req.getParameter("id"));
-            System.out.println(id);
-            productService.delete(id);
-            resp.sendRedirect("/products");
-        }catch (Exception e){
-            String errorMessage = "Your product has not been deleted! Please try again";
-            PageGenerator pageGenerator = PageGenerator.instance();
-            Map<String, Object> parameters = Map.of("errorMessage", errorMessage);
-            String page = pageGenerator.getPage("deleteProduct.html", parameters);
-            resp.getWriter().write(page);
+        if (securityService.isAuth(req, userTokens)) {
+            try {
+                int id = Integer.parseInt(req.getParameter("id"));
+                System.out.println(id);
+                productService.delete(id);
+                resp.sendRedirect("/products");
+            } catch (Exception e) {
+                String errorMessage = "Your product has not been deleted! Please try again";
+                PageGenerator pageGenerator = PageGenerator.instance();
+                Map<String, Object> parameters = Map.of("errorMessage", errorMessage);
+                String page = pageGenerator.getPage("deleteProduct.html", parameters);
+                resp.getWriter().write(page);
+            }
+        } else {
+            resp.sendRedirect("/login");
         }
     }
 }
